Define Compiler type in hardhat/types/solidity
