+++ v-next/hardhat/src/internal/builtin-plugins/solidity/type-extensions.ts
@@ -1,7 +1,6 @@
-import type { Compiler } from "./build-system/compiler/compiler.js";
 import type { SolcConfig } from "../../../types/config.js";
 import type { SolidityBuildSystem } from "../../../types/solidity/build-system.js";
-import type { CompilerInput, CompilerOutput } from "../../../types/solidity.js";
+import type { Compiler, CompilerInput, CompilerOutput } from "../../../types/solidity.js";
 
 import "../../../types/config.js";
 declare module "../../../types/config.js" {
